+++ muduo/base/FileUtil.h
@@ -73,7 +73,7 @@ class AppendFile : noncopyable
 
   void flush();
 
-  size_t writtenBytes() const { return writtenBytes_; }
+  off_t writtenBytes() const { return writtenBytes_; }
 
  private:
 
@@ -81,7 +81,7 @@ class AppendFile : noncopyable
 
   FILE* fp_;
   char buffer_[64*1024];
-  size_t writtenBytes_;
+  off_t writtenBytes_;
 };
 }
